multimodal/requirements.txt
albumentations
aniso8601==9.0.1
asttokens==2.0.5
autofaiss==2.13.1
backcall==0.2.0
braceexpand==0.1.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
clip-anytorch==2.3.1
clip-retrieval==2.24.10
dataclasses==0.6
decorator==5.1.1
docker-pycreds==0.4.0
embedding-reader==1.2.1
executing==0.8.3
ExifRead-nocycle==3.0.1
faiss-cpu==1.7.2
filelock==3.6.0
fire==0.4.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fsspec==2022.2.0
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
h5py==3.6.0
iconclass
huggingface-hub==0.4.0
idna==3.3
imageio==2.16.1
img2dataset==1.26.0
ipython==8.1.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
networkx==2.7.1
nltk==3.7
numpy
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
regex==2022.3.2
requests==2.27.1
sacremoses==0.0.47
scikit-image==0.19.2
scikit-learn
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
sentry-sdk==1.5.7
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.2.9
tokenizers==0.11.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
traitlets==5.1.1
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.8
wandb==0.12.11
wcwidth==0.2.5
webdataset==0.1.103
Werkzeug==2.0.3
yaspin==2.1.0
fastapi==0.65.2
uvicorn==0.15.0
python-multipart==0.0.5
databases==0.5.5
aiosqlite==0.17.0
ICONCLASS==0.10
aiofiles
transcrypt
